requirements 2.txt
fastapi>=0.68.0
uvicorn>=0.15.0
psutil>=5.9.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
alembic>=1.7.4
redis>=4.1.0
psycopg2-binary>=2.9.1
python-multipart>=0.0.5
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
Pillow>=9.0.0  # For icon generation
numpy>=1.21.0  # For icon generation
